requirements.txt
requests==2.32.4
beautifulsoup4==4.13.4
python-telegram-bot==22.2
python-dotenv==1.1.1
